RL-POOL: Renaming and Refactoring MLP Envmodel for Robotinhtm
