+++ examples/rl/Howto 16 - (RL) Model Based Reinforcement Learning.py
@@ -23,7 +23,7 @@ from mlpro.rl.models import *
 from mlpro.rl.pool.envs.robotinhtm import RobotHTM
 from stable_baselines3 import PPO
 from mlpro.wrappers.sb3 import WrPolicySB32MLPro
-from mlpro.rl.pool.envmodels.mlpenvmodel import MLPEnvModel
+from mlpro.rl.pool.envmodels.mlp_robotinhtm import MLPEnvModel
 
 class SimulatedTraining(RLTraining):
     C_NAME = "Simulated"
